+++ controllers/index.js
@@ -1,11 +1,9 @@
 
 const recipe = require("./recipe");
-const search = require("./serch")
 const { getShoppingList } = require("./shoppingList");
 
 module.exports = {
   getShoppingList,
     recipe,
-    search
 
 };
