+++ controllers/recipe/index.js
@@ -2,8 +2,10 @@ const { ctrlWrapper } = require("../../middlewares");
 
 const { getRecipeById } = require("./getRecipeById");
 const { getPopularRecipes } = require("./getPopularRecipes");
+const {searchRecipe} = require("./searchRecipe")
 
 module.exports = {
   getRecipeById: ctrlWrapper(getRecipeById),
   getPopularRecipes: ctrlWrapper(getPopularRecipes),
+  searchRecipe: ctrlWrapper(searchRecipe)
 };
